backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.3
pydantic==2.7.4
redis==5.0.1
httpx==0.25.2
google-generativeai==0.8.2
python-dotenv==1.0.1
